requirements.txt
streamlit>=1.28.0
flask>=2.3.0
flask-cors>=4.0.0
wikipedia-api>=0.6.0
wikipedia>=1.4.0
langchain>=0.0.350
openai>=1.3.0
tensorflow>=2.13.0
transformers>=4.35.0
sentence-transformers>=2.2.2
torch>=2.1.0
datasets>=2.14.0
accelerate>=0.24.0
requests>=2.31.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
plotly>=5.17.0
seaborn>=0.12.0
